Support PAX/POSIX tar on import
POSIX/PAX headers in TAR may be larger than 1 TAR block
(otherwise we see the Pax global headers as a "file" in backwards compatible
mode).

Calling Tar_unix.get_next_header would already read the appropriate amount from
a FD to find the first true file header.
However an imported file can be either compressed or not, and we need to retry
with decompression if parsing as a Tar failed. But we might be importing from a
socket or pipe, so we cannot just seek back to the beginning,
and we cannot use Tar_unix.get_next_header either because we don't have access
to its internal buffer.

Implement a custom Tar header reader using the functor, that first feeds the
block that we already read to the reader, and then the file itself directly
(at which point we turn off the retry because we no longer buffer all that we
read).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>