XSI-2093: CA-422229: Host should not register all IPs to DNS server
Fixes: cba2f1d5e4340b352f32effeacb023a0315d671b
During fix the localhost name issue, An problem was found
Hosts in a pool can not resovle each other with static IP
Thus, an enhancement is applied to push host name and IPs to DNS,
This pushed all IPs of the host into DNS server, including the
storage interface.

This commit just revert the DNS change.
Regarding the resovle issue with static IP, it better goes to
somewhere else like network event hook, or system deamon, if we
do care about it and want a fix.

Signed-off-by: Lin Liu <lin.liu01@citrix.com>